api/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
redis==5.0.7
django-cors-headers==4.4.0
gunicorn==22.0.0